+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/lib/src/types/bitmap.dart
@@ -157,8 +157,22 @@ class BitmapDescriptor {
 
   /// Creates a BitmapDescriptor using an array of bytes that must be encoded
   /// as PNG.
-  static BitmapDescriptor fromBytes(Uint8List byteData) {
-    return BitmapDescriptor._(<Object>[_fromBytes, byteData]);
+  /// On the web, the [size] parameter represents the *physical size* of the
+  /// bitmap, regardless of the actual resolution of the encoded PNG.
+  /// This helps the browser to render High-DPI images at the correct size.
+  /// `size` is not required (and ignored, if passed) in other platforms.
+  static BitmapDescriptor fromBytes(Uint8List byteData, {Size? size}) {
+    assert(byteData.isNotEmpty,
+        'Cannot create BitmapDescriptor with empty byteData');
+    return BitmapDescriptor._(<Object>[
+      _fromBytes,
+      byteData,
+      if (kIsWeb && size != null)
+        <Object>[
+          size.width,
+          size.height,
+        ]
+    ]);
   }
 
   final Object _json;