+++ packages/video_player/video_player_platform_interface/lib/video_player_platform_interface.dart
@@ -6,8 +6,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-import 'method_channel_video_player.dart';
-
 /// The interface that implementations of video_player must implement.
 ///
 /// Platform implementations should extend this class rather than implement it as `video_player`
@@ -21,11 +19,12 @@ abstract class VideoPlayerPlatform extends PlatformInterface {
 
   static final Object _token = Object();
 
-  static VideoPlayerPlatform _instance = MethodChannelVideoPlayer();
+  static VideoPlayerPlatform _instance = _PlaceholderImplementation();
 
-  /// The default instance of [VideoPlayerPlatform] to use.
+  /// The instance of [VideoPlayerPlatform] to use.
   ///
-  /// Defaults to [MethodChannelVideoPlayer].
+  /// Defaults to a placeholder that does not override any methods, and thus
+  /// throws `UnimplementedError` in most cases.
   static VideoPlayerPlatform get instance => _instance;
 
   /// Platform-specific plugins should override this with their own
@@ -112,9 +111,10 @@ abstract class VideoPlayerPlatform extends PlatformInterface {
   }
 }
 
+class _PlaceholderImplementation extends VideoPlayerPlatform {}
+
 /// The video resolution, the number of distinct pixels in each dimension that
 /// can be displayed
-@immutable
 class Resolution {
   /// Creates a [Resolution] with the given [width] and [height].
   const Resolution(this.width, this.height);