+++ packages/camera/camera/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum Flutter version to 2.10.
+
 ## 0.10.0+1
 
 * Fixes avoid_redundant_argument_values lint warnings and minor typos.